git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/branches/prism-fastsim@8229 bbc10eb1-c90d-0410-af57-cb519fbb1720

+++ prism/src/simulator/sampler/SamplerBoundedUntilDisc.java
@@ -68,7 +68,7 @@ public class SamplerBoundedUntilDisc extends SamplerBoolean
 		if (valueKnown)
 			return true;
 
-		int pathSize = path.size();
+		long pathSize = path.size();
 		// Upper bound exceeded
 		if (pathSize > ub) {
 			valueKnown = true;
